+++ templates/base.html
@@ -1,7 +1,7 @@
 {% load i18n %}
 {% load static from staticfiles %}
 {% load bootstrap3 %}
-{% load pathagar_permissions %}
+{% load pathagar_common %}
 
 <!DOCTYPE html>
 <html lang="en">
@@ -49,7 +49,7 @@
             <ul class="nav navbar-nav navbar-right">
                 <li class="nav-item"><a data-name="latest" href="{% url "latest" %}">{% trans "Books" %}</a></li>
                 <li class="nav-item"><a data-name="authors" href="{% url "author_list" %}">{% trans "Authors" %}</a></li>
-                <li class="nav-item"><a href="{% url "latest_feed" %}">{% trans "RSS" %}</a></li>
+                <li class="nav-item"><a href="{% rss_url list_by tag q %}">{% trans "RSS" %}</a></li>
 
                 {% if user|can_upload %}
                 <li class="nav-item"><a href="{% url "book_add" %}">{% trans "Upload" %}</a></li>
